fix(lint): single ANSI stripping implementation, shared with pure modules
CSpell flagged "nqry", a character class fragment of the ANSI regex added in
pluginsVersionUtils. Rather than adding a meaningless word to the dictionary,
the regex is removed: ansiColors.ts (which has no vscode import, so pure
modules can use it) now owns the only implementation, utils.stripAnsi
delegates to it, and pluginsVersionUtils re-exports it.
