+++ src/utils.ts
@@ -951,14 +951,9 @@ export async function getGitParentBranch() {
   return null;
 }
 
-const ansiPattern = [
-  "[\\u001B\\u009B][[\\]()#;?]*(?:(?:(?:(?:;[-a-zA-Z\\d\\/#&.:=?%@~_]+)*|[a-zA-Z\\d]+(?:;[-a-zA-Z\\d\\/#&.:=?%@~_]*)*)?\\u0007)",
-  "(?:(?:\\d{1,4}(?:;\\d{0,4})*)?[\\dA-PR-TZcf-nq-uy=><~]))",
-].join("|");
-const ansiRegex = new RegExp(ansiPattern, "g");
-
+// Single implementation, shared with the modules that can not import vscode
 export function stripAnsi(str: string) {
-  return (str || "").replace(ansiRegex, "");
+  return c.stripAnsiCodes(str);
 }
 
 let IS_WEB_VSCODE: boolean | null = null;
